+++ tests/testthat/test_freqlist.R
@@ -181,6 +181,20 @@ test_that("Changing the labels", {
   )
 
   expect_identical(
+    capture.output(summary(freqlist(TAB.na, na.options = "include", labelTranslations = list("Treatment", "Ethan Rocks")))),
+    c(""                                                                 ,
+      ""                                                                 ,
+      "|Treatment |Ethan Rocks | Freq| cumFreq| freqPercent| cumPercent|",
+      "|:---------|:-----------|----:|-------:|-----------:|----------:|",
+      "|A         |Ethan       |   17|      17|       18.89|      18.89|",
+      "|          |Heinzen     |   16|      33|       17.78|      36.67|",
+      "|          |NA          |    3|      36|        3.33|      40.00|",
+      "|B         |Ethan       |   25|      61|       27.78|      67.78|",
+      "|          |Heinzen     |   29|      90|       32.22|     100.00|"
+    )
+  )
+
+  expect_identical(
     capture.output(summary(freqlist(TAB.na, na.options = "include", labelTranslations = c("Treatment", "Ethan Rocks")))),
     c(""                                                                 ,
       ""                                                                 ,